requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
tensorflow==2.12.0
keras==2.12.0
pillow==9.4.0
numpy==1.23.5
opencv-python==4.7.0.72
requests
mediapipe
- e